+++ man-roxygen/ex-mod_growth_death_rate.R
@@ -1,7 +1,7 @@
 if (interactive()) {
   library(shiny)
-  sapply(file.path("R",list.files("R")), source)
-
+  #sapply(file.path("R",list.files("R")), source)
+  #devtools::load_all()
   ui <- fluidPage(
     tagList(
       Covid19Mirai:::golem_add_external_resources(),