+++ tests/testthat/test-get-datahub.R
@@ -50,7 +50,7 @@ test_that("get_timeseries_by_contagion_day_data returns expected headers", {
 
 test_that("add_growth_death_rate returns expected headers", {
   df <- data %>% add_growth_death_rate()
-  expect_true(all(c("growth_factor_3", "growth_factor_5", "growth_factor_7", "lethality_rate") %in% names(df)))
+  expect_true(all(c("growth_factor_3", "growth_factor_14", "growth_factor_7", "lethality_rate") %in% names(df)))
 })
 
 test_that("test for missing dates and today data", {